requirements.txt
speedtest-cli
flask